# docs/swagger.yaml
definitions:
  command.ChangePasswordCommand:
    properties:
      new_password:
        minLength: 8
        type: string
      old_password:
        type: string
    required:
    - new_password
    - old_password
    type: object
  command.CreateUserCommand:
    properties:
      age:
        maximum: 150
        minimum: 0
        type: integer
      email:
        type: string
      name:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - age
    - email
    - name
    - password
    type: object
  command.UpdateUserCommand:
    properties:
      age:
        maximum: 150
        minimum: 0
        type: integer
      email:
        type: string
      name:
        type: string
    required:
    - age
    - email
    - name
    type: object
info:
  contact: {}
paths:
  /health:
    get:
      description: Check if the service is healthy
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "503":
          description: Service Unavailable
          schema:
            additionalProperties: true
            type: object
      summary: Health check
      tags:
      - health
  /metrics:
    get:
      description: Get application metrics
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get metrics
      tags:
      - metrics
  /users:
    get:
      description: Get paginated list of users with optional filters
      parameters:
      - description: Search by name or email
        in: query
        name: search
        type: string
      - description: Minimum age
        in: query
        name: age_min
        type: integer
      - description: Maximum age
        in: query
        name: age_max
        type: integer
      - description: Sort field (id, name, email, age, created_at)
        in: query
        name: sort
        type: string
      - description: Sort order (asc, desc)
        in: query
        name: order
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Users list
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: List users with filters
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Create a new user with name, email, password, and age
      parameters:
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/command.CreateUserCommand'
      produces:
      - application/json
      responses:
        "201":
          description: User created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid input
          schema:
            additionalProperties: true
            type: object
        "409":
          description: User already exists
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Create a new user
      tags:
      - users
  /users/{id}:
    delete:
      description: Delete a user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User deleted
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid user ID
          schema:
            additionalProperties: true
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Delete user
      tags:
      - users
    get:
      description: Get a single user by their ID (with Redis caching)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User found
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid user ID
          schema:
            additionalProperties: true
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get user by ID
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update user information
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/command.UpdateUserCommand'
      produces:
      - application/json
      responses:
        "200":
          description: User updated
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid input
          schema:
            additionalProperties: true
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Email already exists
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Update user
      tags:
      - users
  /users/{id}/change-password:
    put:
      consumes:
      - application/json
      description: Change password for a user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Password data
        in: body
        name: password
        required: true
        schema:
          $ref: '#/definitions/command.ChangePasswordCommand'
      produces:
      - application/json
      responses:
        "200":
          description: Password changed
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid input
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Incorrect old password
          schema:
            additionalProperties: true
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Change user password
      tags:
      - users
  /users/search:
    get:
      description: Search users by keyword
      parameters:
      - description: Search keyword
        in: query
        name: q
        required: true
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Search results
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid input
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Search users
      tags:
      - users
swagger: "2.0"